requirements.txt

crewai
crewai-tools


fastapi
uvicorn[standard]


pydantic
python-dotenv


requests


reportlab

google-generativeai

firebase_admin

